+++ packages/@kagal-taistamp/src/index.ts
@@ -5,10 +5,21 @@ export const VERSION: string = pkg.version;
 
 export * from './const';
 export {
+  composeSignaturePayload,
   newTaistampHandler,
   type TaistampHandlerConfig,
-  taistampSignedPayload,
 } from './handler';
+export {
+  asLeapSeconds,
+  extractLeapSeconds,
+  type LeapSeconds,
+  TAI_LEAP_SECONDS,
+  TAI_LEAP_SECONDS_MAX,
+} from './leap-seconds';
+export {
+  asNonce,
+  type Nonce,
+} from './nonce';
 export {
   newEd25519Signer,
   type Signer,